refactor golang application to differntiate between cli and tui
