+++ cmd/cli/commands/root.go
@@ -1,0 +1,26 @@
+package commands
+
+import (
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+// RootCmd represents the base command when called without any subcommands
+var RootCmd = &cobra.Command{
+	Use:   "dataloaf",
+	Short: "DataLoaf",
+	Long:  ``,
+}
+
+// Execute adds all child commands to the root command and sets flags appropriately.
+func Execute() {
+	err := RootCmd.Execute()
+	if err != nil {
+		os.Exit(1)
+	}
+}
+
+func init() {
+	RootCmd.Flags().BoolP("toggle", "t", false, "Help message for toggle")
+}
