+++ cmd/cli/main.go
@@ -1,0 +1,7 @@
+package main
+
+import command "dataloaf/cli/commands"
+
+func main() {
+	command.Execute()
+}
