--- golang/main.go
@@ -1,10 +1,0 @@
-/*
-Copyright © 2024 NAME HERE <EMAIL ADDRESS>
-*/
-package main
-
-import "dataloaf/cmd"
-
-func main() {
-	cmd.Execute()
-}
